Create dummy error reporter for dev/test
