+++ config/test/django.env
@@ -1,3 +1,4 @@
+DJANGO_SETTINGS_MODULE=osuchan.settings
 SECRET_KEY="generatearandomsecretforproduction"
 DEBUG=True
 ALLOWED_HOSTS=["local.syrin.me"]
@@ -23,3 +24,5 @@ OSU_CLIENT_REDIRECT_URI=fakeredirecturi
 OSU_API_V1_KEY=fakeapikey
 
 DISCORD_WEBHOOK_URL_ERROR_LOG=fakediscordwebhookurl
+
+USE_DUMMY_ERROR_REPORTER=True
